[403]: Move a bunch of reflection-related config stuff into the reflector.
